+++ .idea/workspace.xml
@@ -4,9 +4,7 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="99d44297-2c64-4063-accd-ffdbf8af173f" name="Changes" comment="sprint 3 base &#10;http://localhost:8080/enterprises/&#10;El sistema devuelve responses 200 en la ruta&#10;/enterprises con los siguientes verbos:&#10;GET&#10;POST&#10;El sistema devuelve responses 200 en la ruta&#10;/enterprises/[id] con los siguientes verbos:&#10;GET&#10;PATCH&#10;DELETE&#10;&#10;http://localhost:8080/users/&#10;El sistema devuelve responses 200 en la ruta&#10;/users con los siguientes verbos:&#10;GET&#10;POST&#10;El sistema devuelve responses 200 en la ruta&#10;/users/[id] con los siguientes verbos:&#10;GET&#10;PATCH&#10;DELETE">
-      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-    </list>
+    <list default="true" id="99d44297-2c64-4063-accd-ffdbf8af173f" name="Changes" comment="sprint 3 base &#10;http://localhost:8080/enterprises/&#10;El sistema devuelve responses 200 en la ruta&#10;/enterprises con los siguientes verbos:&#10;GET&#10;POST&#10;El sistema devuelve responses 200 en la ruta&#10;/enterprises/[id] con los siguientes verbos:&#10;GET&#10;PATCH&#10;DELETE&#10;&#10;http://localhost:8080/users/&#10;El sistema devuelve responses 200 en la ruta&#10;/users con los siguientes verbos:&#10;GET&#10;POST&#10;El sistema devuelve responses 200 en la ruta&#10;/users/[id] con los siguientes verbos:&#10;GET&#10;PATCH&#10;DELETE" />
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
@@ -104,10 +102,15 @@
     &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
     &quot;WebServerToolWindowFactoryState&quot;: &quot;false&quot;,
     &quot;last_opened_file_path&quot;: &quot;E:/INTELLIJ IDEA CODIGOS/TrabajoEnClaseCiclo3&quot;,
-    &quot;settings.editor.selected.configurable&quot;: &quot;preferences.pluginManager&quot;,
+    &quot;settings.editor.selected.configurable&quot;: &quot;project.propVCSSupport.DirectoryMappings&quot;,
     &quot;spring.configuration.checksum&quot;: &quot;5d52821ea14dc2245e01c1836aece84e&quot;
   }
 }</component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="E:\INTELLIJ IDEA CODIGOS\SoftCodeSA\SoftCodeSA\src\main\java\com\linag\ciclo3\repositories" />
+    </key>
+  </component>
   <component name="RunDashboard">
     <option name="configurationTypes">
       <set>
@@ -129,11 +132,6 @@
         <option name="Make" enabled="true" />
       </method>
     </configuration>
-    <recent_temporary>
-      <list>
-        <item itemvalue="Spring Boot.main" />
-      </list>
-    </recent_temporary>
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
@@ -153,6 +151,8 @@
       <workItem from="1662463054232" duration="61000" />
       <workItem from="1662463688214" duration="112000" />
       <workItem from="1662584833791" duration="9160000" />
+      <workItem from="1662777491141" duration="164000" />
+      <workItem from="1662777840685" duration="469000" />
     </task>
     <task id="LOCAL-00001" summary="sprint 2 base">
       <created>1662315101269</created>
@@ -235,6 +235,9 @@
     </option>
   </component>
   <component name="VcsManagerConfiguration">
+    <ignored-roots>
+      <path value="$PROJECT_DIR$" />
+    </ignored-roots>
     <MESSAGE value="sprint 2 base" />
     <MESSAGE value="sprint 3 base" />
     <MESSAGE value="sprint 3 base funciona html" />
